Mateo

+++ src/main/java/com/example/stokapp/product/domain/ProductService.java
@@ -1,7 +1,9 @@
 package com.example.stokapp.product.domain;
 
+import com.example.stokapp.auth.AuthImpl;
 import com.example.stokapp.configuration.DtoConfig;
 import com.example.stokapp.exceptions.NotFound;
+import com.example.stokapp.exceptions.UnauthorizeOperationException;
 import com.example.stokapp.product.infrastructure.ProductRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,12 +22,25 @@ public class ProductService {
 
     @Autowired
     ModelMapper mapper;
+    @Autowired
+    private AuthImpl authImpl;
 
     //ADD PRODUCT
-    public void addProduct(Product product) { productRepository.save(product); }
+    public void addProduct(Product product) {
+        String username = authImpl.getCurrentEmail();
+        if(username == null) {
+            throw new UnauthorizeOperationException("Not allowed");
+        }
+        productRepository.save(product);
+    }
 
     //DELETE PRODUCT
     public void deleteProduct(Long productId) {
+        String username = authImpl.getCurrentEmail();
+        if(username == null) {
+            throw new UnauthorizeOperationException("Not allowed");
+        }
+
         Product product = productRepository.findById(productId)
                 .orElseThrow(() -> new NotFound("Product not found"));
         productRepository.delete(product);
@@ -33,6 +48,11 @@ public class ProductService {
 
     //UPDATE PRODUCT
     public void updateProduct(Long productId, Product updatedProduct) {
+        String username = authImpl.getCurrentEmail();
+        if(username == null) {
+            throw new UnauthorizeOperationException("Not allowed");
+        }
+
         Product existingProduct = productRepository.findById(productId)
                 .orElseThrow(() -> new RuntimeException("Product not found"));
 
@@ -44,18 +64,13 @@ public class ProductService {
         productRepository.save(existingProduct);
     }
 
-    //BUSCAR PRODUCTO POR nombre
-    public ProductDto getProductByName(String nombre) {
-        Product products = productRepository.findByName(nombre)
-                .orElseThrow(() -> new NotFound("Product not found"));
-        ProductDto productDto = mapper.map(products, ProductDto.class);
-
-        return productDto;
-    }
-
-
     // BUSCAR TODOS LOS PRODUCTOS
     public List<Product> getAllProducts() {
+        String username = authImpl.getCurrentEmail();
+        if(username == null) {
+            throw new UnauthorizeOperationException("Not allowed");
+        }
+
         List<Product> products = productRepository.findAll();
         if(products.isEmpty()){
             throw new RuntimeException("No products found.");
